[Properties] Correctly manage volume and play speed
Added a TransportPlaySpeeds property.  This property is not
defined in the MPRIS specification, but is required to easily set rate value.

The volume and playspeed parameters are now properly implemented and are
consistent with the MPRIS specification.

Player interface hasn't got a max_volume property since MPRIS
specifications only ever expose volumes in a well known [0.0…1.0] range.
This attribute is needed to map [0…max-volume] to MPRIS volume range.

Signed-off-by: Sébastien Bianti <sebastien.bianti@linux.intel.com>